pyproject.toml
[project]
name = "newbest-mqtt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "paho-mqtt>=2.1.0",
]
